Grooming: Separate distro to different files
