+++ daisy_workflows/image_test/configuration/centos.py
@@ -1,0 +1,31 @@
+#!/usr/bin/env python2
+# Copyright 2018 Google Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import abc
+
+import redhat
+
+
+class CentOSTests(redhat.RedHatTests):
+  __metaclass__ = abc.ABCMeta
+  pass
+
+
+class CentOS6Tests(redhat.RedHat6Tests):
+  pass
+
+
+class CentOS7Tests(redhat.RedHat7Tests):
+  pass
